fix: use resume instead of sessionId for session continuation and add shutdown cleanup
The continueSession resume path incorrectly passed sessionId to the SDK query options,
which is meant for new sessions. Now passes resume: sessionId to properly load conversation
history. Added a scope finalizer to abort all active SDK connections and update DB status
on process shutdown. Extracted bootstrap and background fiber logic into internal/ modules.
