+++ whatever-code/src/claude/orchestrator.ts
@@ -1,10 +1,9 @@
-import {
-  query,
-  type Options as QueryOptions,
-  type SDKMessage,
-  type SDKUserMessage,
+import type {
+  Options as QueryOptions,
+  SDKMessage,
+  SDKUserMessage,
 } from "@anthropic-ai/claude-agent-sdk";
-import { Effect, Queue, Stream } from "effect";
+import { Effect, Queue } from "effect";
 import { v7 } from "uuid";
 import { ulid } from "ulid";
 import { ClaudeChatError } from "../api/definitions/claude.js";
@@ -13,14 +12,13 @@ import {
   claudeSessionSqliteEntity,
   claudeTurnSqliteEntity,
 } from "../db/claude.js";
-import { ContinueSessionParams, QueryParams } from "./schema.js";
-
-interface ActiveSession {
-  abortController: AbortController;
-  inputQueue: Queue.Queue<SDKUserMessage>;
-  outputQueue: Queue.Queue<SDKMessage>;
-  turnId: string;
-}
+import {
+  ContinueSessionParams,
+  QueryParams,
+  type ActiveSession,
+} from "./schema.js";
+import { recoverInterruptedSessions } from "./internal/recover-interrupted.js";
+import { startBackgroundFiber } from "./internal/background-fiber.js";
 
 const makeUserMessage = (
   sessionId: string,
@@ -32,127 +30,67 @@ const makeUserMessage = (
   session_id: sessionId,
 });
 
+const persistNewTurn = (
+  sessionId: string,
+  turnId: string,
+  userMessage: SDKUserMessage,
+) =>
+  Effect.all([
+    claudeSessionSqliteEntity.update(
+      { id: sessionId },
+      { status: "in_progress" },
+    ),
+    claudeTurnSqliteEntity.insert({
+      id: turnId,
+      sessionId,
+      status: "in_progress",
+      init: null,
+      result: null,
+    }),
+    claudeMessageSqliteEntity.insert({
+      id: ulid(),
+      sessionId,
+      turnId,
+      data: userMessage,
+    }),
+  ]).pipe(Effect.orDie);
+
 export class ClaudeOrchestrator extends Effect.Service<ClaudeOrchestrator>()(
   "ClaudeOrchestrator",
   {
     scoped: Effect.gen(function* () {
       const activeSessions = new Map<string, ActiveSession>();
 
-      // Bootstrap: mark all in_progress sessions and their turns as interrupted
-      const { items: inProgressSessions } = yield* claudeSessionSqliteEntity
-        .query("byStatus", { pk: { status: "in_progress" }, sk: { ">": null } })
-        .pipe(Effect.orDie);
+      yield* recoverInterruptedSessions;
 
-      yield* Effect.all(
-        inProgressSessions.map(({ value }) =>
-          claudeSessionSqliteEntity
-            .update({ id: value.id }, { status: "interrupted" })
-            .pipe(Effect.orDie),
-        ),
-        { discard: true },
-      );
-
-      yield* Effect.all(
-        inProgressSessions.map(({ value }) =>
-          claudeTurnSqliteEntity
-            .query("bySession", {
-              pk: { sessionId: value.id },
-              sk: { ">": null },
-            })
-            .pipe(
-              Effect.flatMap(({ items }) =>
-                Effect.all(
-                  items
-                    .filter(({ value: turn }) => turn.status === "in_progress")
-                    .map(({ value: turn }) =>
-                      claudeTurnSqliteEntity
-                        .update({ id: turn.id }, { status: "interrupted" })
-                        .pipe(Effect.orDie),
-                    ),
-                  { discard: true },
-                ),
+      yield* Effect.addFinalizer(() =>
+        Effect.forEach(
+          activeSessions.entries(),
+          ([sessionId, session]) =>
+            Effect.all(
+              [
+                claudeTurnSqliteEntity
+                  .update({ id: session.turnId }, { status: "interrupted" })
+                  .pipe(Effect.orDie),
+                claudeSessionSqliteEntity
+                  .update({ id: sessionId }, { status: "interrupted" })
+                  .pipe(Effect.orDie),
+              ],
+              { discard: true },
+            ).pipe(
+              Effect.ensuring(
+                Effect.sync(() => session.abortController.abort()),
               ),
-              Effect.orDie,
             ),
+          { discard: true },
         ),
-        { discard: true },
       );
 
-      const startBackgroundFiber = (
-        sessionId: string,
-        session: ActiveSession,
-        options?: QueryOptions,
-      ) => {
-        const queryResult = query({
-          prompt: Stream.toAsyncIterable(Stream.fromQueue(session.inputQueue)),
-          options: {
-            ...options,
-            sessionId,
-            abortController: session.abortController,
-          },
-        });
-
-        return Effect.forkScoped(
-          Stream.fromAsyncIterable(
-            queryResult,
-            (e) => new ClaudeChatError({ message: String(e) }),
-          ).pipe(
-            Stream.runForEach((message) =>
-              Effect.gen(function* () {
-                const turnId = session.turnId;
-
-                yield* claudeMessageSqliteEntity
-                  .insert({ id: ulid(), sessionId, turnId, data: message })
-                  .pipe(Effect.orDie);
-
-                yield* Queue.offer(session.outputQueue, message);
-
-                if (message.type === "system" && message.subtype === "init") {
-                  yield* claudeTurnSqliteEntity
-                    .update({ id: turnId }, { init: message })
-                    .pipe(Effect.orDie);
-                } else if (message.type === "result") {
-                  const result = message;
-                  const status = result.is_error ? "error" : "success";
-                  yield* claudeTurnSqliteEntity
-                    .update({ id: turnId }, { status, result })
-                    .pipe(Effect.orDie);
-                  yield* claudeSessionSqliteEntity
-                    .update({ id: sessionId }, { status })
-                    .pipe(Effect.orDie);
-                }
-              }),
-            ),
-            Effect.catchAll(() =>
-              activeSessions.has(sessionId)
-                ? Effect.all(
-                    [
-                      claudeTurnSqliteEntity
-                        .update({ id: session.turnId }, { status: "error" })
-                        .pipe(Effect.orDie),
-                      claudeSessionSqliteEntity
-                        .update({ id: sessionId }, { status: "error" })
-                        .pipe(Effect.orDie),
-                    ],
-                    { discard: true },
-                  )
-                : Effect.void,
-            ),
-            Effect.ensuring(
-              Effect.gen(function* () {
-                yield* Queue.shutdown(session.outputQueue);
-                activeSessions.delete(sessionId);
-              }),
-            ),
-          ),
-        );
-      };
-
-      const startNewSession = (
+      const initSession = (
         sessionId: string,
         turnId: string,
         userMessage: SDKUserMessage,
-        options?: QueryOptions,
+        queryOptions?: QueryOptions,
       ) =>
         Effect.gen(function* () {
           const inputQueue = yield* Queue.unbounded<SDKUserMessage>();
@@ -165,7 +103,12 @@ export class ClaudeOrchestrator extends Effect.Service<ClaudeOrchestrator>()(
           };
           activeSessions.set(sessionId, session);
           yield* Queue.offer(inputQueue, userMessage);
-          yield* startBackgroundFiber(sessionId, session, options);
+          yield* startBackgroundFiber(
+            activeSessions,
+            sessionId,
+            session,
+            queryOptions,
+          );
         });
 
       const createSession = (params: typeof QueryParams.Type) =>
@@ -173,6 +116,7 @@ export class ClaudeOrchestrator extends Effect.Service<ClaudeOrchestrator>()(
           const sessionId = v7();
           const turnId = ulid();
           const userMessage = makeUserMessage(sessionId, params.prompt);
+
           yield* Effect.all([
             claudeSessionSqliteEntity.insert({
               id: sessionId,
@@ -194,34 +138,12 @@ export class ClaudeOrchestrator extends Effect.Service<ClaudeOrchestrator>()(
             }),
           ]).pipe(Effect.orDie);
 
-          yield* startNewSession(sessionId, turnId, userMessage, params.options);
+          yield* initSession(sessionId, turnId, userMessage, {
+            ...params.options,
+            sessionId,
+          });
           return { sessionId };
         });
-
-      const persistNewTurn = (
-        sessionId: string,
-        turnId: string,
-        userMessage: SDKUserMessage,
-      ) =>
-        Effect.all([
-          claudeSessionSqliteEntity.update(
-            { id: sessionId },
-            { status: "in_progress" },
-          ),
-          claudeTurnSqliteEntity.insert({
-            id: turnId,
-            sessionId,
-            status: "in_progress",
-            init: null,
-            result: null,
-          }),
-          claudeMessageSqliteEntity.insert({
-            id: ulid(),
-            sessionId,
-            turnId,
-            data: userMessage,
-          }),
-        ]).pipe(Effect.orDie);
 
       const continueSession = (params: typeof ContinueSessionParams.Type) =>
         Effect.gen(function* () {
@@ -238,47 +160,40 @@ export class ClaudeOrchestrator extends Effect.Service<ClaudeOrchestrator>()(
               );
             }
 
-            const newTurnQueue = yield* Queue.unbounded<SDKMessage>();
-
             yield* persistNewTurn(params.sessionId, turnId, userMessage);
-
             existing.turnId = turnId;
-            existing.outputQueue = newTurnQueue;
-
+            existing.outputQueue = yield* Queue.unbounded<SDKMessage>();
             yield* Queue.offer(existing.inputQueue, userMessage);
             return;
           }
 
-          // Inactive session — resume
           yield* persistNewTurn(params.sessionId, turnId, userMessage);
-          yield* startNewSession(
-            params.sessionId,
-            turnId,
-            userMessage,
-            params.options,
-          );
+          yield* initSession(params.sessionId, turnId, userMessage, {
+            ...params.options,
+            resume: params.sessionId,
+          });
         });
 
       const stopSession = (sessionId: string) =>
         Effect.gen(function* () {
           const session = activeSessions.get(sessionId);
-          if (session) {
-            activeSessions.delete(sessionId);
-            yield* Effect.all(
-              [
-                claudeTurnSqliteEntity
-                  .update({ id: session.turnId }, { status: "interrupted" })
-                  .pipe(Effect.orDie),
-                claudeSessionSqliteEntity
-                  .update({ id: sessionId }, { status: "interrupted" })
-                  .pipe(Effect.orDie),
-              ],
-              { discard: true },
-            );
-            session.abortController.abort();
-            yield* Queue.shutdown(session.inputQueue);
-            yield* Queue.shutdown(session.outputQueue);
-          }
+          if (!session) return;
+
+          activeSessions.delete(sessionId);
+          yield* Effect.all(
+            [
+              claudeTurnSqliteEntity
+                .update({ id: session.turnId }, { status: "interrupted" })
+                .pipe(Effect.orDie),
+              claudeSessionSqliteEntity
+                .update({ id: sessionId }, { status: "interrupted" })
+                .pipe(Effect.orDie),
+            ],
+            { discard: true },
+          );
+          session.abortController.abort();
+          yield* Queue.shutdown(session.inputQueue);
+          yield* Queue.shutdown(session.outputQueue);
         });
 
       return { createSession, continueSession, stopSession };
